Replanteamiento de la DB.

+++ src/main/java/gt/sat/aduanas/gec/repositories/AdDetalleRazonPeticionMGRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdDetalleRazonPeticionMGRepository extends CrudRepository<AdDetalleRazonPeticionMG, Long> {
+public interface AdDetalleRazonPeticionMGRepository extends CrudRepository<AdDetalleRazonPeticionMG, Integer> {
 }
